"ads2_m7_a1_1: 6/6 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ Assignments/m7/1. MST/Solution.java
@@ -7,11 +7,11 @@ public class Solution {
 		EdgeWeightedGraph edgGraph = new EdgeWeightedGraph(ver);
 		while (scan.hasNext()) {
 		String[] arr = scan.nextLine().split(" ");
-		Edge edge = new Edge(Integer.parseInt(arr[0]), Integer.parseInt(arr[1]), Integer.parseInt(arr[2]));
+		Edge edge = new Edge(Integer.parseInt(arr[0]), Integer.parseInt(arr[1]), Double.parseDouble(arr[2]));
 		edgGraph.addEdge(edge);
 	}
 	LazyPrimMST prim = new LazyPrimMST(edgGraph);
-	System.out.printf("%.4f\n", prim.weight());
+	System.out.printf("%.5f\n", prim.weight());
 
 	}
 }
